requirements_Version1.txt
fastapi==0.112.2
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==2.1.1
requests==2.32.3
openai==1.47.0
pytest==8.3.2